delete org tests
